Auth middleware
Cookie created in app.js, then a middleware applied to every route makes sure that you're logged in or redirects you to home (and will console.log important stuff).
Edits made to users.js were so that the hashedPassword is never stored in the cookie, as we should never need that there.